refactor(_core): extract WORKFLOW data into shared workflow_data.py + Sprint 4 brief
The WORKFLOW dict + _ROBUSTNESS_OVERRIDES + _ROBUSTNESS_WORKFLOW_LEVELS
+ Transition dataclass now live in megaplan/_core/workflow_data.py.
megaplan/_core/workflow.py re-exports them (back-compat); the Pipeline
in megaplan/_pipeline/planning.py imports from the shared module.

Single source of truth — both runtimes read from the same definitions.
Sprint 4 (briefs/sprint-4-elegance.md) inverts the relationship: the
Pipeline becomes the source and the legacy dicts get deleted.
